Fix module level documentation
