+++ crates/rune/src/modules/hash.rs
@@ -1,11 +1,11 @@
-//! The `std::hash` module.
+//! Hashing types.
 
 use crate as rune;
 use crate::runtime::Hasher;
 use crate::{ContextError, Module};
 
+/// Hashing types.
 #[rune::module(::std::hash)]
-/// Types for dealing with hashing in Rune.
 pub fn module() -> Result<Module, ContextError> {
     #[allow(unused_mut)]
     let mut module = Module::from_meta(self::module_meta)?;
