+++ crates/rune/src/modules/num.rs
@@ -1,12 +1,17 @@
-//! The `std::num` module.
+//! Working with numbers.
 
 use core::num::{ParseFloatError, ParseIntError};
 
+use crate as rune;
 use crate::{ContextError, Module};
 
-/// Install the core package into the given functions namespace.
+/// Working with numbers.
+///
+/// This module provides types generic for working over numbers, such as errors
+/// when a number cannot be parsed.
+#[rune::module(::std::num)]
 pub fn module() -> Result<Module, ContextError> {
-    let mut module = Module::with_crate_item("std", ["num"])?;
+    let mut module = Module::from_meta(self::module_meta)?;
     module.ty::<ParseFloatError>()?;
     module.ty::<ParseIntError>()?;
     Ok(module)
