+++ crates/rune/src/modules/stream.rs
@@ -1,11 +1,13 @@
-//! The `std::stream` module.
+//! Asynchronous streams.
 
+use crate as rune;
 use crate::runtime::{Stream, Vm};
 use crate::{ContextError, Module};
 
-/// Construct the `std::stream` module.
+/// Asynchronous streams.
+#[rune::module(::std::stream)]
 pub fn module() -> Result<Module, ContextError> {
-    let mut module = Module::with_crate_item("std", ["stream"])?;
+    let mut module = Module::from_meta(self::module_meta)?;
     module.ty::<Stream<Vm>>()?;
     module.associated_function("next", Stream::<Vm>::next_shared)?;
     module.associated_function("resume", Stream::<Vm>::resume_shared)?;
